+++ effects-cats/shared/src/main/scala/busymachines/pureharm/effects/package.scala
@@ -25,27 +25,6 @@ package object effects
   type AttemptT[F[_], R] = cats.data.EitherT[F, Throwable, R]
   val AttemptT: cats.data.EitherT.type = cats.data.EitherT
 
-  /** Useful since we don't have partial kind application by default
-    * Usage:
-    * {{{
-    *   def canFail[F[_]: ApplicativeAttempt, T](p1: T) : F[T] = ???
-    * }}}
-    */
-  @scala.deprecated("Use ApplicativeThrow instead", "0.1.0")
-  type ApplicativeAttempt[F[_]] = cats.ApplicativeError[F, Throwable]
-
-  /** Useful since we don't have partial kind application by default
-    * Usage:
-    * {{{
-    *   def canFail[F[_]: MonadAttempt, T](p1: T) : F[T] = ???
-    * }}}
-    */
-  @scala.deprecated("Use MonadThrow instead", "0.1.0")
-  type MonadAttempt[F[_]] = cats.MonadError[F, Throwable]
-
-  @scala.deprecated("Use BracketThrow instead", "0.1.0")
-  type BracketAttempt[F[_]] = cats.effect.Bracket[F, Throwable]
-
   /** !!! N.B. !!!
     * NEVER, EVER wildcard import this, AND, cats.implicits, or anything from the cats packages.
     *